make get last n more readable

+++ components/consensusmanager/src/session.rs
@@ -498,7 +498,7 @@ impl ConsensusSessionOwned {
     pub async fn async_intrusive_pruning_point_update(&self, new_pruning_point: Hash, syncer_sink: Hash) -> ConsensusResult<()> {
         self.clone().spawn_blocking(move |c| c.intrusive_pruning_point_update(new_pruning_point, syncer_sink)).await
     }
-    pub async fn async_get_n_last_pruning_points(&self, n: u64) -> Vec<Hash> {
+    pub async fn async_get_n_last_pruning_points(&self, n: usize) -> Vec<Hash> {
         self.clone().spawn_blocking(move |c| c.get_n_last_pruning_points(n)).await
     }
 }
